requirements.txt
opencv-python~=4.10.0.84
setuptools~=69.5.1
numpy~=1.24.3
torchvision~=0.18.1
pillow~=10.4.0
requests~=2.32.2
psutil~=6.0.0
pandas~=2.2.2
matplotlib~=3.8.3
pyyaml~=6.0.1
tqdm~=4.66.4
scipy~=1.13.1